// src/models/job.ts
export interface Job {
  address: string;
  category: string;
  date_posted: string;
  description: string;
  experience: string;
  job_id: string;
  lat: string;
  lng: string;
  name: string;
  recruiter_id_fk: string;
  salary: string;
  salary_frequency: string;
  skills: string;
  type: string;
}